ci(workflow): add GitHub Actions workflow for building and pushing Docker images
Create a new GitHub Actions workflow to automate the build and push
process of Docker images. This workflow triggers on pushes to the main
branch, tags, and pull requests, ensuring that Docker images are built
and pushed to the GitHub Container Registry.

build(Dockerfile): add Dockerfile for building OpenTelemetry Collector

Introduce a Dockerfile to build a custom OpenTelemetry Collector with
Solace Receiver. This Dockerfile uses a multi-stage build to compile
the collector and package it into a final image based on the
OpenTelemetry Collector Contrib image. This setup facilitates
consistent and reproducible builds for deployment.