ci(test.yml): update test paths to focus on internal/receiver directory for more targeted testing
This change narrows the scope of the tests to the internal/receiver directory, ensuring that only relevant tests are run, which can improve test efficiency and focus.

build(Dockerfile): download Solace C API archive during build process

This change automates the download of the Solace C API archive, removing the need for a pre-existing local copy and ensuring the latest version is used during the build.